#50 refactor: actor の store/context を _stores/actor/ へ統合
- actor-settings と同じ構成 (types.ts/constants.ts/store.ts/context.tsx/index.ts) に揃える
- context.tsx は共通生成関数 createStoreContext 経由に置き換え
