+++ src/prototypes/time-control/time-control-03/_components/current-position-marker/index.tsx
@@ -1,11 +1,10 @@
 import { memo, useEffect, useRef } from 'react'
 
 import { getTickMs } from '../../../time-control-02/_lib/get-tick-ms'
-import { useActorStore } from '../../_contexts/actor-store-context'
 import { usePositionStoreApi } from '../../_contexts/position-store-context'
 import { useStageTransform } from '../../_contexts/stage-transform-context'
 import { toPixelStyle } from '../../_lib/stage-coords'
-import { DEFAULT_ACTOR_INFO } from '../../_stores/actor-store'
+import { DEFAULT_ACTOR_INFO, useActorStore } from '../../_stores/actor'
 import { DEFAULT_POSITION } from '../../_stores/position-store'
 import { ActorId, Position } from '../../types'
