+++ src/prototypes/time-control/time-control-03/_lib/build-schedule.ts
@@ -1,5 +1,5 @@
 import { getTickMs } from '../../time-control-02/_lib/get-tick-ms'
-import { ActorInfo, DEFAULT_ACTOR_INFO } from '../_stores/actor-store'
+import { ActorInfo, DEFAULT_ACTOR_INFO } from '../_stores/actor'
 import { ActorId, MovePath, ScheduleRow } from '../types'
 
 /**
@@ -12,7 +12,7 @@ import { ActorId, MovePath, ScheduleRow } from '../types'
  *
  * @param actorIds 対象 actor 一覧
  * @param pathById actor ごとの残り経路 (`_stores/path-store.ts`)
- * @param actorById actor ごとの移動速度・tick 発生頻度 (`_stores/actor-store.ts`)
+ * @param actorById actor ごとの移動速度・tick 発生頻度 (`_stores/actor/store.ts`)
  * @param commonGameTimeMs 共通ゲームクロックの現在値 (`_stores/game-clock-store.ts`)
  */
 export const buildSchedule = (
