+++ src/prototypes/time-control/time-control-03/_stores/actor/context.tsx
@@ -1,0 +1,9 @@
+import { createStoreContext } from '@/stores/utils/create-store-context'
+
+import { ActorState } from './types'
+
+const { StoreContext, useStoreSelector } =
+  createStoreContext<ActorState>('Actor')
+
+export const ActorStoreContext = StoreContext
+export const useActorStore = useStoreSelector
